adding changes for the all-flag and format flag
