+++ crates/nu-command/src/commands/mod.rs
@@ -81,8 +81,6 @@ mod tests {
             whole_stream_command(StrContains),
             whole_stream_command(StrIndexOf),
             whole_stream_command(StrTrim),
-            whole_stream_command(StrTrimLeft),
-            whole_stream_command(StrTrimRight),
             whole_stream_command(StrStartsWith),
             whole_stream_command(StrEndsWith),
             //whole_stream_command(StrCollect),
